simplify column export
