+++ src/pydiverse/transform/_internal/backend/polars.py
@@ -6,7 +6,7 @@ from uuid import UUID
 import polars as pl
 
 from pydiverse.transform._internal.backend.table_impl import TableImpl
-from pydiverse.transform._internal.backend.targets import Pandas, Polars, Target
+from pydiverse.transform._internal.backend.targets import Polars, Target
 from pydiverse.transform._internal.ops import ops
 from pydiverse.transform._internal.ops.op import Ftype
 from pydiverse.transform._internal.tree import types, verbs
@@ -70,15 +70,6 @@ class PolarsImpl(TableImpl):
                 lf = lf.collect()
             lf.name = nd.name
             return lf
-
-        raise AssertionError
-
-    @staticmethod
-    def export_col(expr: ColExpr, target: Target) -> pl.Series:
-        if isinstance(target, Polars):
-            ...
-        elif isinstance(target, Pandas):
-            ...
 
         raise AssertionError
