+++ src/pydiverse/transform/_internal/backend/table_impl.py
@@ -14,7 +14,7 @@ from pydiverse.transform._internal.errors import NotSupportedError
 from pydiverse.transform._internal.ops import ops
 from pydiverse.transform._internal.ops.op import Ftype
 from pydiverse.transform._internal.tree.ast import AstNode
-from pydiverse.transform._internal.tree.col_expr import Col, ColExpr
+from pydiverse.transform._internal.tree.col_expr import Col
 from pydiverse.transform._internal.tree.types import Dtype
 
 try:
@@ -121,9 +121,6 @@ class TableImpl(AstNode):
     ) -> Any: ...
 
     @classmethod
-    def export_col(cls, expr: ColExpr, target: Target): ...
-
-    @classmethod
     def get_impl(cls, op: Operator, sig: Sequence[Dtype]) -> Any:
         if (impl := cls.impl_store.get_impl(op, sig)) is not None:
             return impl
@@ -138,6 +135,15 @@ class TableImpl(AstNode):
                 f"operation `{op.name}` is not supported by the backend "
                 f"`{cls.__name__.lower()[:-4]}`"
             ) from err
+
+
+def get_backend(nd: AstNode) -> type[TableImpl]:
+    from pydiverse.transform._internal.pipe.verbs import Verb
+
+    if isinstance(nd, Verb):
+        return get_backend(nd.child)
+    assert isinstance(nd, TableImpl) and nd is not TableImpl
+    return nd.__class__
 
 
 with TableImpl.impl_store.impl_manager as impl:
