+++ tests/test_polars_table.py
@@ -580,20 +580,21 @@ class TestPolarsLazyImpl:
             check_row_order=False,
         )
 
-    def test_col_export(self, tbl1, tbl2):
-        assert_equal(df1.get_column("col1"), tbl1.col1 >> export(Polars()))
+    def test_col_export(self, tbl1: pdt.Table, tbl2: pdt.Table):
+        assert_equal(df1.get_column("col1"), tbl1.col1.export(Polars()))
         t = tbl2 >> mutate(u=tbl2.col1 * tbl2.col2, v=-tbl2.col3)
         t_ex: pl.DataFrame = t >> export(Polars(lazy=False))
         assert_equal(
             (t_ex["u"] + t_ex["col2"]).exp() - t_ex["v"],
-            ((t.u + C.col2).exp() - t.v) >> export(Polars()),
+            (t >> mutate(h=(t.u + C.col2).exp() - t.v)).h.export(Polars()),
         )
         e = t >> inner_join(
             tbl1, tbl1.col1.cast(pdt.Float64()) <= tbl2.col1 + tbl2.col3
         )
         e_ex = e >> export(Polars(lazy=False))
         assert_equal(
-            e_ex["col2"] + e_ex["col1_df1"], (e.col2 + tbl1.col1) >> export(Polars())
+            e_ex["col2"] + e_ex["col1_df1"],
+            (e >> mutate(j=e.col2 + tbl1.col1)).j.export(Polars()),
         )
